Task #4588: COmmit merge

+++ branches/Cobalt-Task4588/RTCP/Cobalt/GPUProc/src/cuda/Kernels/Kernel.h
@@ -45,11 +45,11 @@ namespace LOFAR
       struct Parameters
       {
         Parameters(const Parset& ps);
-        size_t nrStations;
-        size_t nrChannelsPerSubband;
-        size_t nrSamplesPerChannel;
-        size_t nrSamplesPerSubband;
-        size_t nrPolarizations;
+        unsigned nrStations;
+        unsigned nrChannelsPerSubband;
+        unsigned nrSamplesPerChannel;
+        unsigned nrSamplesPerSubband;
+        unsigned nrPolarizations;
         bool dumpBuffers;
         std::string dumpFilePattern;
       };
@@ -85,7 +85,7 @@ namespace LOFAR
       // Explicit destructor, because the implicitly generated one is public.
       ~Kernel();
       
-      const size_t maxThreadsPerBlock;
+      const unsigned maxThreadsPerBlock;
       gpu::Grid globalWorkSize;
       gpu::Block localWorkSize;
       size_t nrOperations, nrBytesRead, nrBytesWritten;
